pre codeblock horizontal scroll
This allows divs containing preformatted codeblocks to scroll horizontally. For an example of the non-scrolling current behavior, look at the shell script on [this note page](https://digital-garden-jekyll-template.netlify.app/your-first-note) with browser <1100px. The div containing a preformatted block is not resizing with the grid, which is preventing the backlinks box from being visible. This is also a problem on mobile devices, since horizontal scrolling of the body is not allowed.

This fix, `display: grid`, allows the div.highlight element to resize along with the grid container, fixing both the responsive backlink box issue and horizontal scrolling of the codeblock for browser widths <600px.